+++ src/index.ts
@@ -253,13 +253,13 @@ app.get("/sitemap.xml", async (c) => {
     (_, i) =>
       `<sitemap><loc>https://api.egdata.app/sitemap.xml?page=${
         i + 1
-      }</loc><lastmod>${new Date().toISOString()}</lastmod></sitemap>`,
+      }</loc></sitemap>`,
   ).join("")}
 </sitemapindex>`;
 
     return c.text(siteMapIndex, 200, {
       "Content-Type": "application/xml",
-      "Cache-Control": `max-age=${cacheTimeInSec}, stale-while-revalidate=${cacheStaleTimeInSec}`,
+      "Cache-Control": `public, max-age=${cacheTimeInSec}, stale-while-revalidate=${cacheStaleTimeInSec}`,
     });
   }
 
@@ -301,7 +301,7 @@ app.get("/sitemap.xml", async (c) => {
 
   return c.text(siteMap, 200, {
     "Content-Type": "application/xml",
-    "Cache-Control": `max-age=${cacheTimeInSec}, stale-while-revalidate=${cacheStaleTimeInSec}`,
+    "Cache-Control": `public, max-age=${cacheTimeInSec}, stale-while-revalidate=${cacheStaleTimeInSec}`,
   });
 });
 
@@ -1441,26 +1441,18 @@ app.get("/items-sitemap.xml", async (c) => {
     (_, i) =>
       `<sitemap><loc>https://api.egdata.app/items-sitemap.xml?page=${
         i + 1
-      }</loc><lastmod>${new Date().toISOString()}</lastmod></sitemap>`,
+      }</loc></sitemap>`,
   ).join("")}
 </sitemapindex>`;
 
     return c.text(siteMapIndex, 200, {
       "Content-Type": "application/xml",
-      "Cache-Control": `max-age=${cacheTimeInSec}, stale-while-revalidate=${cacheStaleTimeInSec}`,
+      "Cache-Control": `public, max-age=${cacheTimeInSec}, stale-while-revalidate=${cacheStaleTimeInSec}`,
     });
   }
 
   // Generate individual sitemap page
-  const sections = [
-    "offers",
-    "assets",
-    "achievements",
-    "related",
-    "metadata",
-    "changelog",
-    "media",
-  ];
+  const sections = [] as const;
 
   const items = await Item.find(
     {},
@@ -1498,7 +1490,7 @@ app.get("/items-sitemap.xml", async (c) => {
 
   return c.text(siteMap, 200, {
     "Content-Type": "application/xml",
-    "Cache-Control": `max-age=${cacheTimeInSec}, stale-while-revalidate=${cacheStaleTimeInSec}`,
+    "Cache-Control": `public, max-age=${cacheTimeInSec}, stale-while-revalidate=${cacheStaleTimeInSec}`,
   });
 });